Adjust Folder Format Move modules rsTimedLoop and ArduinoController to custom module folder. Split readArduino and singleLegControl to seperate cmake builds
