+++ modules/custom/rsTimedLoop/rsTimedLoop.cpp
@@ -1,4 +1,4 @@
-#include "rs_timed_loop.h"
+#include "rsTimedLoop.h"
 #include "iostream"
 
 using namespace std;
